% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{check_metadata_names}
\alias{check_metadata_names}
\title{check_metadata_names}
\usage{
check_metadata_names(col_df, df_name = "", df_type = NULL)
}
\arguments{
\item{col_df}{a character with colnames of df}

\item{df_name}{a name of dataframe ("" by default)}

\item{df_type}{a type of a dataframe (NULL by default)}
}
\value{
a charvec with corrected colnames of df
}
\description{
Check whether all metadata names are correct
}
\examples{
 td1 <- get_test_data1()
 m_file <- td1$m_file
 m_data <- readxl::read_excel(m_file)
 result <- check_metadata_names(col_df = colnames(m_data))

}
