% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_file_format}
\alias{detect_file_format}
\title{Detect formar of results data}
\usage{
detect_file_format(results_file)
}
\arguments{
\item{results_file}{path to results data}
}
\value{
string with the information about detected data format of results
}
\description{
Detect formar of results data
}
