% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{.check_file_structure}
\alias{.check_file_structure}
\title{Check the structure of raw data}
\usage{
.check_file_structure(
  df,
  iB,
  iF,
  iS,
  results_filename,
  readout_offset,
  n_row,
  n_col,
  barcode_col
)
}
\description{
Check the structure of raw data
}
\keyword{internal}
