% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pset_to_gdrDF.R
\name{.createPseudoData}
\alias{.createPseudoData}
\title{Add in pseudo-data for duration and cell reference division time}
\usage{
.createPseudoData(dt)
}
\value{
data.table in proper data format
}
\description{
Add in pseudo-data for duration and cell reference division time
}
\keyword{internal}
