% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pset_to_gdrDF.R
\name{.extractDoseResponse}
\alias{.extractDoseResponse}
\title{Get dose and viability readouts and melt into large data table}
\usage{
.extractDoseResponse(pset)
}
\value{
data.table with dose and viability readouts
}
\description{
Get dose and viability readouts and melt into large data table
}
\keyword{internal}
