% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{.get_plate_size}
\alias{.get_plate_size}
\title{Get plate size}
\usage{
.get_plate_size(df)
}
\value{
charvec with plate dims
}
\description{
Get plate size
}
\details{
All plate sizes assume 1.5x nrows = ncolumns.
}
\keyword{internal}
