% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pset_to_gdrDF.R
\name{.removeNegatives}
\alias{.removeNegatives}
\title{Remove negative viabilities}
\usage{
.removeNegatives(dataset)
}
\value{
data.table with positive values in column \code{ReadoutValue}
}
\description{
Remove negative viabilities
}
\keyword{internal}
