% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{.standardize_untreated_values}
\alias{.standardize_untreated_values}
\title{Standardize untreated values to ignore cases}
\usage{
.standardize_untreated_values(df)
}
\value{
data.frame with standardized untreated values
}
\description{
Standardize untreated values to ignore cases
}
\keyword{internal}
