% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\docType{class}
\name{gdr_test_data}
\alias{gdr_test_data}
\title{gDR Test Data object}
\value{
object class \code{gdr_test_data} with primary test data
}
\description{
Object class \code{gdr_test_data} is build by function \code{\link[=get_test_data]{get_test_data()}}
}
\section{Slots}{

\describe{
\item{\code{manifest_path}}{character, path to manifest file}

\item{\code{result_path}}{character, path(s) to results file}

\item{\code{template_path}}{character, path(s) to data.frame with template data}

\item{\code{ref_m_df}}{character, data.frame with manifest data}

\item{\code{ref_r1_r2}}{character, path to reference file with raw data for treated & untreated}

\item{\code{ref_r1}}{character, path to reference file with raw data for treated}

\item{\code{ref_t1_t2}}{character, path to reference template file with treated & untreated data}

\item{\code{ref_t1}}{character, path to reference template file with treated data}
}}

