% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRimport-package.R
\docType{package}
\name{gDRimport-package}
\alias{gDRimport}
\alias{gDRimport-package}
\title{gDRimport: Package for handling the import of dose-response data}
\value{
package help page
}
\description{
The package is a part of the gDR suite. It helps to prepare raw drug response data for downstream processing. It mainly contains helper functions for importing/loading/validating dose-response data provided in different file formats.
}
\note{
To learn more about functions start with \code{help(package = "gDRimport")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDRimport}
  \item \url{https://gdrplatform.github.io/gDRimport/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDRimport/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Sergiu Mocanu
  \item Dariusz Scigocki
  \item Allison Vuong
  \item Luca Gerosa (\href{https://orcid.org/0000-0001-6805-9410}{ORCID})
  \item Janina Smola (\href{https://orcid.org/0009-0007-4347-7748}{ORCID})
}

}
\keyword{internal}
