% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{get_df_from_raw_edited_EnVision_df}
\alias{get_df_from_raw_edited_EnVision_df}
\title{Get final results (as a data.table)
from raw edited EnVision data.table}
\usage{
get_df_from_raw_edited_EnVision_df(
  df,
  barcode_idx,
  barcode_col,
  n_row,
  n_col,
  fname,
  sheet_name,
  headers
)
}
\arguments{
\item{df}{raw data.table}

\item{barcode_idx}{numeric vector with barcode indices}

\item{barcode_col}{column number for barcode data}

\item{n_row}{number of rows}

\item{n_col}{number of columns}

\item{fname}{file name}

\item{sheet_name}{name of the Excel sheet}

\item{headers}{list with the headersa}
}
\value{
data.table derived from EnVision data
}
\description{
Get final results (as a data.table)
from raw edited EnVision data.table
}
\keyword{load_files}
