% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{get_df_from_raw_unedited_EnVision_df}
\alias{get_df_from_raw_unedited_EnVision_df}
\title{Get final results (as a data.frame)
from raw unedited EnVision data.frame}
\usage{
get_df_from_raw_unedited_EnVision_df(df, n_row, n_col, barcode_col)
}
\arguments{
\item{df}{raw data.frame}

\item{n_row}{number of rows}

\item{n_col}{number of columns}

\item{barcode_col}{column number for barcode data}
}
\description{
Get final results (as a data.frame)
from raw unedited EnVision data.frame
}
