% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{get_EnVision_properties}
\alias{get_EnVision_properties}
\title{Get properties of EnVision data}
\usage{
get_EnVision_properties(results.list, fname)
}
\arguments{
\item{results.list}{list with EnVision data}

\item{fname}{name of the input file}
}
\value{
list with EnVision propertiesa
}
\description{
This function return properties of EnVision data
}
\keyword{internal}
\keyword{load_files}
