% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{get_exception_data}
\alias{get_exception_data}
\title{get exception data}
\usage{
get_exception_data(status_code = NULL)
}
\arguments{
\item{status_code}{A numeric value}
}
\value{
A data.table row with exception data or all exceptions
}
\description{
get exception data
}
\examples{
get_exception_data(1)
get_exception_data()
}
\keyword{correction_exception}
