% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{get_expected_template_sheets}
\alias{get_expected_template_sheets}
\title{Get names of the sheets expected in templates xlsx}
\usage{
get_expected_template_sheets(type = c("all", "core", "optional"))
}
\arguments{
\item{type}{charvec type of the sheets}
}
\value{
string with type of the sheets
}
\description{
Get names of the sheets expected in templates xlsx
}
