% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_data4}
\alias{get_test_data4}
\title{get test data3 (EnVision)}
\usage{
get_test_data4()
}
\value{
list with with input data (manifest/template/result paths)
and related reference data (RDS file paths)
}
\description{
get test data3 (EnVision)
}
\examples{
get_test_data4()
}
