% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_Tecan_data}
\alias{get_test_Tecan_data}
\title{get test Tecan data}
\usage{
get_test_Tecan_data()
}
\value{
list with with input data (manifest/template/result paths)
and related reference data (qs2 file paths)
}
\description{
get test Tecan data
}
\examples{
get_test_Tecan_data()
}
\keyword{test_data}
