% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_tsv_data}
\alias{get_test_tsv_data}
\title{get test tsv data}
\usage{
get_test_tsv_data()
}
\value{
list with with input data (manifest/template/result paths)
and related reference data (.qs2 file paths)
}
\description{
get test tsv data
}
\examples{
get_test_tsv_data()
}
\keyword{test_data}
