% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(
  manifest_file,
  df_template_files,
  results_file,
  instrument = "EnVision"
)
}
\arguments{
\item{manifest_file}{character, file path(s) to manifest(s)}

\item{df_template_files}{data.frame, with datapaths and names of results file(s)
or character with file path of templates file(s)}

\item{results_file}{data.frame, with datapaths and names of results file(s)
or character with file path of results file(s)}

\item{instrument}{character}
}
\description{
This functions loads and checks the data file(s)
}
\examples{
 td1 <- get_test_data1()
 l_tbl <- load_data(td1$m_file, td1$t_files, td1$r_files)

}
