% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_templates_xlsx}
\alias{load_templates_xlsx}
\title{Load templates from xlsx}
\usage{
load_templates_xlsx(template_file, template_filename = NULL)
}
\arguments{
\item{template_file}{character, file path(s) to template(s)}

\item{template_filename}{character, file name(s)}
}
\description{
This functions loads and checks the template file(s)
}
