% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\docType{methods}
\name{manifest_path}
\alias{manifest_path}
\alias{manifest_path,gdr_test_data-method}
\title{Method manifest_path}
\usage{
manifest_path(x)

\S4method{manifest_path}{gdr_test_data}(x)
}
\arguments{
\item{x}{object class gdr_test_data}
}
\value{
value of slot \code{manifest_path}
}
\description{
Method for object gdr_test_data - access to slot \code{manifest_path}
}
\examples{
 td <- get_test_data()
 manifest_file_path <- manifest_path(td)

}
\keyword{internal}
\keyword{methods}
\keyword{test_data_class}
