% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\name{manifest_path}
\alias{manifest_path}
\title{Method manifest_path}
\usage{
manifest_path(x)
}
\arguments{
\item{x}{object class gdr_test_data}
}
\description{
Method for object gdr_test_data - access to slot \code{manifest_path}
}
