% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{mgrepl}
\alias{mgrepl}
\title{grep wrapper to support multiple patterns}
\usage{
mgrepl(patterns, x, do_unlist = TRUE, ...)
}
\arguments{
\item{patterns}{charvec with patterns to be checked}

\item{x}{charvec with data}

\item{do_unlist}{logical_flag unlist the final results?}

\item{...}{additional argument}
}
\value{
list of charvec with grep output
}
\description{
grep wrapper to support multiple patterns
}
\keyword{correction_exception}
\keyword{internal}
