% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doses.R
\name{parse_D300}
\alias{parse_D300}
\title{Parse D300}
\usage{
parse_D300(D300_file)
}
\arguments{
\item{D300_file}{string, file path to D300 .tdd file}
}
\description{
This function parses a D300 *.tdd file (XML format) into a data.frame
}
