% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_EnVision_xlsx}
\alias{read_EnVision_xlsx}
\title{Read in single xlsx data from EnVision}
\usage{
read_EnVision_xlsx(results_file, results_sheet)
}
\arguments{
\item{results_file}{character, file path(s) to results file(s)}

\item{results_sheet}{results sheet names}
}
\value{
data.table with results data
}
\description{
Read in single xlsx data from EnVision
}
\keyword{internal}
\keyword{load_files}
