% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_excel_to_dt}
\alias{read_excel_to_dt}
\title{Read excel file and transorm it into data.table object}
\usage{
read_excel_to_dt(path, ...)
}
\arguments{
\item{path}{path to excel file}

\item{...}{other arguments that should be passed into \code{readxl::read_excel}}
}
\value{
data.table object with read excel file
}
\description{
Read excel file and transorm it into data.table object
}
\examples{
datasets <- readxl::readxl_example("datasets.xlsx")
read_excel_to_dt(datasets)
}
