% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_in_EnVision_file}
\alias{read_in_EnVision_file}
\title{Read EnVision file}
\usage{
read_in_EnVision_file(file, nrows, seps)
}
\arguments{
\item{file}{input file from EnVision}

\item{nrows}{maximum number of file rows to be processed}

\item{seps}{potential field separators of the input file}
}
\value{
list with one element per EnVisoin input file
}
\description{
This function reads file from the EnVision Workstation
}
\keyword{internal}
\keyword{load_files}
