% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_in_manifest_file}
\alias{read_in_manifest_file}
\title{read manifest files}
\usage{
read_in_manifest_file(manifest_file, available_formats)
}
\arguments{
\item{manifest_file}{character, file path(s) to manifest(s)}

\item{available_formats}{charvec with available file formats}
}
\value{
a data.table with manifest data
}
\description{
read manifest files
}
\keyword{internal}
\keyword{load_files}
