% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_in_result_files}
\alias{read_in_result_files}
\title{Read in results files}
\usage{
read_in_result_files(results_file, results_filename, headers)
}
\arguments{
\item{results_file}{data.table, with datapaths and names of results file(s)
or character with file path of results file(s)}

\item{results_filename}{charvect with file names}

\item{headers}{list of headers identified in the result files}
}
\value{
data.table with results data
}
\description{
Read in results files
}
\keyword{internal}
\keyword{load_files}
