% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_in_template_xlsx}
\alias{read_in_template_xlsx}
\title{Read in xlsx template files}
\usage{
read_in_template_xlsx(template_file, template_filename, template_sheets)
}
\arguments{
\item{template_file}{character, file path(s) to template(s)}

\item{template_filename}{character, file name(s)}

\item{template_sheets}{template sheet names}
}
\value{
data.table with templates data
}
\description{
Read in xlsx template files
}
\keyword{internal}
\keyword{load_files}
