% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\name{result_path}
\alias{result_path}
\title{Method result_path}
\usage{
result_path(x)
}
\arguments{
\item{x}{object class gdr_test_data}
}
\description{
Method for object gdr_test_data - access to slot \code{result_path}
}
