% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D300.R
\name{save_drug_info_per_well}
\alias{save_drug_info_per_well}
\title{for each drug create a Gnumber and Concentration information for each well}
\usage{
save_drug_info_per_well(trt_info, trt_gnumber_conc, wb, idfs)
}
\description{
for each drug create a Gnumber and Concentration information for each well
}
