% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_file_type_info}
\alias{save_file_type_info}
\title{save_file_type_info}
\usage{
save_file_type_info(
  v,
  save_dir,
  normKeysFileName = "normKeys.json",
  dfRawDataFileName = "dfRawData.tsv",
  fileTypeInfoName = "fileTypeInfo.csv"
)
}
\arguments{
\item{v}{a list with filenames}

\item{save_dir}{an output directory}

\item{normKeysFileName}{a filename with normalization keys ('normKeys.json' by default)}

\item{dfRawDataFileName}{a filename with rawdata ('dfRawData.tsv' by default)}

\item{fileTypeInfoName}{a filename with file type information ('fileTypeInfo.csv' by default)}
}
\value{
\code{NULL} invisibly.
}
\description{
Save information about file types
}
