% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_df}
\alias{standardize_df}
\title{standardize_df}
\usage{
standardize_df(df)
}
\arguments{
\item{df}{data.table or data.frame for standardization}
}
\value{
a standardized data.table
}
\description{
Transform all the columns in a data.table to character type, and transform
data.frame to data.table
}
\examples{
standardize_df(iris)

}
