% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_record_values}
\alias{standardize_record_values}
\title{standardize_record_values}
\usage{
standardize_record_values(x, dictionary = DICTIONARY)
}
\arguments{
\item{x}{a named array}

\item{dictionary}{a named array}
}
\value{
a named array with updated names
}
\description{
map values to a dictionary
}
