% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\name{template_path}
\alias{template_path}
\title{Method template_path}
\usage{
template_path(x)
}
\arguments{
\item{x}{object class gdr_test_data}
}
\description{
Method for object gdr_test_data - access to slot \code{template_path}
}
