% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{validate_template_xlsx}
\alias{validate_template_xlsx}
\title{Validate template xlsx data}
\usage{
validate_template_xlsx(template_file, template_filename, template_sheets, idx)
}
\arguments{
\item{template_file}{character, file path(s) to template(s)}

\item{template_filename}{character, file name(s)}

\item{template_sheets}{template sheet names}

\item{idx}{template file index}
}
\value{
\code{NULL} invisibly.
}
\description{
Validate template xlsx data
}
\keyword{internal}
\keyword{load_files}
