% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_ref_data_df}
\alias{write_ref_data_df}
\title{write_ref_data_df}
\usage{
write_ref_data_df(lData, outDir, prefix = "ref")
}
\arguments{
\item{lData}{a list with dataset}

\item{outDir}{an output directory}

\item{prefix}{a prefix of reference file names ('ref' by default)}
}
\value{

}
\description{
Write reference dataframe
}
