% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_ref_data_se}
\alias{write_ref_data_se}
\title{write_ref_data_se}
\usage{
write_ref_data_se(se, outDir, prefix = "ref")
}
\arguments{
\item{se}{a SummarizedExperiment with DR data}

\item{outDir}{an output directory}

\item{prefix}{a prefix of reference file names ('ref' by default)}
}
\value{

}
\description{
write_ref_data_se
}
