% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{are_template_sheets_valid}
\alias{are_template_sheets_valid}
\title{are template sheet valid?}
\usage{
are_template_sheets_valid(ts)
}
\arguments{
\item{ts}{list with (per file) template sheets}
}
\value{
logical flag
}
\description{
are template sheet valid?
}
\seealso{
get_xl_sheets
}
