% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{check_metadata_against_spaces}
\alias{check_metadata_against_spaces}
\title{Check metadata against spaces}
\usage{
check_metadata_against_spaces(corrected_names, df_name)
}
\arguments{
\item{corrected_names}{a charvec with corrected colnames of df}

\item{df_name}{a name of dataframe ("" by default)}
}
\value{
a charvec with corrected colnames of df
}
\description{
Check metadata against spaces
}
