% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{check_metadata_field_names}
\alias{check_metadata_field_names}
\title{Check metadata field names}
\usage{
check_metadata_field_names(corrected_names, df_name)
}
\arguments{
\item{corrected_names}{a charvec with corrected colnames of df}

\item{df_name}{a name of data.table ("" by default)}
}
\value{
a charvec with corrected colnames of df
}
\description{
Check metadata field names
}
