% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{check_metadata_headers}
\alias{check_metadata_headers}
\title{Check whether metadata headers are correct and make fixes if needed}
\usage{
check_metadata_headers(corrected_names, df_name)
}
\arguments{
\item{corrected_names}{a charvec with corrected colnames of df}

\item{df_name}{a name of data.table ("" by default)}
}
\value{
a charvec with corrected colnames of df
}
\description{
Check whether metadata headers are correct and make fixes if needed
}
