% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{check_metadata_names}
\alias{check_metadata_names}
\title{check_metadata_names}
\usage{
check_metadata_names(col_df, df_name = "", df_type = NULL)
}
\arguments{
\item{col_df}{a character with colnames of df}

\item{df_name}{a name of data.table ("" by default)}

\item{df_type}{a type of a data.table (NULL by default)}
}
\value{
a charvec with corrected colnames of df
}
\description{
Check whether all metadata names are correct
}
\examples{
 td <- get_test_data()
 m_file <- manifest_path(td)
 m_data <- read_excel_to_dt(m_file)
 result <- check_metadata_names(col_df = colnames(m_data))

}
