% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{check_metadata_req_col_names}
\alias{check_metadata_req_col_names}
\title{Check metadata for required column names}
\usage{
check_metadata_req_col_names(col_df, df_name, df_type)
}
\arguments{
\item{col_df}{a charvec with corrected colnames of df}

\item{df_name}{a name of dataframe ("" by default)}

\item{df_type}{a type of a dataframe (NULL by default)}
}
\description{
Check metadata for required column names
}
