% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE_to_PSet.R
\name{convert_MAE_to_PSet}
\alias{convert_MAE_to_PSet}
\title{Convert MultiAssayExperiment to TreatmentResponseExperiment}
\usage{
convert_MAE_to_PSet(mae, pset_name)
}
\arguments{
\item{mae}{A MultiAssayExperiment object generated by gDR.}

\item{pset_name}{A character string specifying the name of the resulting PharmacoSet object.}
}
\value{
A PharmacoSet object.
}
\description{
This function converts a MultiAssayExperiment generated by gDR into a TreatmentResponseExperiment for use in the PharmacoGx package. The resulting PharmacoSet can be used for pharmacogenomic analysis of drug response.
}
\examples{
# Convert a MultiAssayExperiment object to a PharmacoSet object
pset <- convert_MAE_to_PSet(mae, "my_pset")

}
