% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{correct_template_sheets}
\alias{correct_template_sheets}
\title{Correct names of the template sheets (if required)}
\usage{
correct_template_sheets(tfiles)
}
\arguments{
\item{tfiles}{charvec with paths to template files}
}
\description{
Correct names of the template sheets (if required)
}
