% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_file_format}
\alias{detect_file_format}
\title{Detect format of results data}
\usage{
detect_file_format(results_file)
}
\arguments{
\item{results_file}{path to results data}
}
\value{
string of the detected file format
}
\description{
Detect format of results data
}
\examples{
td2 <- get_test_Tecan_data()
detect_file_format(td2$r_files[1])

}
