% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{.check_against_single_template_sheet}
\alias{.check_against_single_template_sheet}
\title{Evaluate if template file with single sheet is present,
if the name of the sheet is correct and if it can be fixed}
\usage{
.check_against_single_template_sheet(ts)
}
\arguments{
\item{ts}{list with template sheets info}
}
\description{
get sheets for given set of XLS files
}
