% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{.check_file_structure}
\alias{.check_file_structure}
\title{Check the structure of raw data}
\usage{
.check_file_structure(
  df,
  filename,
  sheet_name,
  readout_offset,
  n_row,
  n_col,
  bcode_idx,
  bcode_col
)
}
\value{
\code{NULL} invisibly.
}
\description{
Check the structure of raw data
}
\keyword{internal}
