% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{.fill_empty_wells}
\alias{.fill_empty_wells}
\title{Correct plates with not fully filled readout values}
\usage{
.fill_empty_wells(
  df,
  plate_rows,
  data_rows,
  exp_row,
  exp_col,
  numeric_regex = "^\\\\d+$"
)
}
\description{
Correct plates with not fully filled readout values
}
\keyword{internal}
