% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{enhance_raw_edited_EnVision_df}
\alias{enhance_raw_edited_EnVision_df}
\title{Enhance raw edited EnVision data.table}
\usage{
enhance_raw_edited_EnVision_df(df, barcode_col, headers)
}
\arguments{
\item{df}{raw data.table}

\item{barcode_col}{column number for barcode data}

\item{headers}{list with the headers}
}
\value{
data.table derived from EnVision data
}
\description{
Enhance raw edited EnVision data.table
}
\keyword{load_files}
