% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRimport-package.R
\docType{package}
\name{gDRimport-package}
\alias{gDRimport}
\alias{gDRimport-package}
\title{gDRimport: Package for handling import of dose-response data}
\value{
package help page
}
\description{
Package is a part of the gDR suite. It helps to prepare raw drug response data for downstream processing. It mainly contains helper functions for importing/loading/validating dose response data provided in different file formats.
}
\note{
To learn more about functions start with \code{help(package = "gDRimport")}
}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com}

Authors:
\itemize{
  \item Bartosz Czech
  \item Marc Hafner
  \item Sergiu Mocanu
  \item Dariusz Scigocki
  \item Allison Vuong
  \item Luca Gerosa
  \item Janina Smola
}

}
\keyword{internal}
