% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{get_envision_properties}
\alias{get_envision_properties}
\title{Get properties of EnVision data}
\usage{
get_envision_properties(results.list)
}
\arguments{
\item{results.list}{list with EnVision data}
}
\description{
This function return properties of EnVision data
}
