% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{get_plate_info_from_template_xlsx}
\alias{get_plate_info_from_template_xlsx}
\title{Get plate info from template xlsx}
\usage{
get_plate_info_from_template_xlsx(template_file, Gnumber_idx, idx)
}
\arguments{
\item{template_file}{character, file path(s) to template(s)}

\item{Gnumber_idx}{index with Gnumber data}

\item{idx}{template file index}
}
\value{
list with plate info
}
\description{
Get plate info from template xlsx
}
