% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_D300_data}
\alias{get_test_D300_data}
\title{get test D300 data}
\usage{
get_test_D300_data()
}
\value{
list with with input data (manifest/template/result paths)
and related reference data (qs file paths)
}
\description{
get test D300 data
}
\examples{
get_test_D300_data()
}
\keyword{test_data}
