% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_data}
\alias{get_test_data}
\title{get primary test data}
\usage{
get_test_data()
}
\value{
object class "gdr_test_data" with with input data (manifest/template/result paths)
and related reference data (qs file paths)
}
\description{
get primary test data
}
\examples{
get_test_data()

}
