% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_data2}
\alias{get_test_data2}
\title{get test data2 (Tecan)
list with with input data (manifest/template/result paths)
and related reference data (RDS file paths)}
\usage{
get_test_data2()
}
\description{
get test data2 (Tecan)
list with with input data (manifest/template/result paths)
and related reference data (RDS file paths)
}
\examples{
get_test_data2()
}
