% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_EnVision_data}
\alias{get_test_EnVision_data}
\title{get test EnVision data}
\usage{
get_test_EnVision_data()
}
\value{
list with with input data (manifest/template/result paths)
and related reference data (.qs file paths)
}
\description{
get test EnVision data
}
\examples{
get_test_EnVision_data()
}
