% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{get_xl_sheets}
\alias{get_xl_sheets}
\title{XLS/XLSX #
get sheets for given set of XLS files}
\usage{
get_xl_sheets(files)
}
\arguments{
\item{files}{charvec with file paths}
}
\description{
get sheets for given set of XLS files
}
