% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pset_to_gdrDF.R
\name{getPSet}
\alias{getPSet}
\title{Get PharmacoSet}
\usage{
getPSet(pset_name, psetDir = getwd(), canonical = FALSE, timeout = 600)
}
\value{
PharmacoSet object
}
\description{
Get PharmacoSet
}
\examples{
suppressMessages(getPSet("Tavor_2020", psetDir = system.file("extdata/pset", package = "gDRimport")))

}
