% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doses.R
\name{import_D300}
\alias{import_D300}
\title{Import D300}
\usage{
import_D300(D300_file, Gnum_file, destination_path)
}
\arguments{
\item{D300_file}{character, file path to D300 file}

\item{Gnum_file}{character, file path to file with mapping from D300 names to Gnumbers}

\item{destination_path}{character, path to folder where template 
files will be generated}

\item{instrument}{character}
}
\description{
This functions takes a D300 file and generates corresponding template files
}
