% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_utils.R
\name{is_readable_v}
\alias{is_readable_v}
\title{is_readable_v
Check if all paths in vector are readable}
\usage{
is_readable_v(paths)
}
\arguments{
\item{paths}{a character with path(s)}
}
\value{

}
\description{
is_readable_v
Check if all paths in vector are readable
}
