% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_manifest}
\alias{load_manifest}
\title{Load manifest}
\usage{
load_manifest(manifest_file)
}
\arguments{
\item{manifest_file}{character, file path(s) to manifest(s)}
}
\value{
list with manifest data.table and headers
}
\description{
This functions loads and checks the manifest file(s)
}
\examples{
 td <- get_test_data()
 ml <- load_manifest(manifest_path(td))

}
