% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_results}
\alias{load_results}
\title{Load results}
\usage{
load_results(df_results_files, instrument = "EnVision")
}
\arguments{
\item{df_results_files}{data.frame, with datapaths and names of results file(s)
or character with file path of results file(s)}

\item{instrument}{character}
}
\description{
This functions loads and checks the results file(s)
}
