% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_results_EnVision}
\alias{load_results_EnVision}
\title{Load envision results from xlsx}
\usage{
load_results_EnVision(results_file, headers = gDRutils::get_env_identifiers())
}
\arguments{
\item{results_file}{character, file path(s) to results file(s)}

\item{headers}{list of headers identified in the manifest}
}
\description{
This functions loads and checks the results file(s)
}
