% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_results_EnVision}
\alias{load_results_EnVision}
\title{Load results from xlsx}
\usage{
load_results_EnVision(results_file)
}
\arguments{
\item{results_file}{character, file path(s) to template(s)}
}
\description{
This functions loads and checks the results file(s)
}
