% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_results_EnVision_new}
\alias{load_results_EnVision_new}
\title{Load results from EnVision_new (CSV)}
\usage{
load_results_EnVision_new(
  results_file,
  headers = gDRutils::get_env_identifiers()
)
}
\arguments{
\item{results_file}{character, file path(s) to result file(s)}

\item{headers}{list of headers identified in the manifest}
}
\value{
data.table with results data
}
\description{
This functions loads and checks the results file(s) from a Revvity EnVision_new instrument
in the CSV format.
}
\keyword{load_files}
