% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_results_Incucyte}
\alias{load_results_Incucyte}
\title{Load incucyte results from plain text}
\usage{
load_results_Incucyte(results_file, headers = gDRutils::get_env_identifiers())
}
\arguments{
\item{results_file}{list of strings: file paths to result paths from individual plates}

\item{headers}{list of headers identified in the manifest}
}
\value{
data.table derived from Incucyte data
}
\description{
This functions loads incucyte time-course cell count file
}
\keyword{load_files}
