% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D300.R
\name{parse_D300_xml}
\alias{parse_D300_xml}
\title{Parse D300}
\usage{
parse_D300_xml(D300_file)
}
\arguments{
\item{D300_file}{string, file path to D300 .tdd file}
}
\value{
data.table representing input \code{D300_file}.
}
\description{
This function parses a D300 *.tdd file (XML format) into a data.table
}
\examples{
td3 <- get_test_D300_data()
fs <- td3[["f_96w"]]
dose_df <- parse_D300_xml(fs[["d300"]])
}
