% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_EnVision}
\alias{read_EnVision}
\title{Read envision files}
\usage{
read_EnVision(file, nrows = 10000, seps = c(",", "\\t"))
}
\arguments{
\item{file}{}

\item{nrows}{}

\item{seps}{}
}
\value{

}
\description{
This function reads file from the EnVision Workstation
}
