% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_in_template_sheet_xlsx}
\alias{read_in_template_sheet_xlsx}
\title{Read in data from xlsx template sheet}
\usage{
read_in_template_sheet_xlsx(template_file, template_sheets, idx, plate_info)
}
\arguments{
\item{template_file}{character, file path(s) to template(s)}

\item{template_sheets}{template sheet names}

\item{idx}{template file index}

\item{plate_info}{list with plate info}
}
\value{
data.table with template data
}
\description{
Read in data from xlsx template sheet
}
