% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{read_in_tsv_template_files}
\alias{read_in_tsv_template_files}
\title{read in tsv template files}
\usage{
read_in_tsv_template_files(template_file)
}
\arguments{
\item{template_file}{character, file path(s) to template(s)}
}
\description{
read in tsv template files
}
