% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ref_data}
\alias{read_ref_data}
\title{read_ref_data}
\usage{
read_ref_data(inDir, prefix = "ref")
}
\arguments{
\item{inDir}{a directory path of reference data}

\item{prefix}{a prefix of reference file names ('ref' by default)}
}
\value{
a list of reference data
}
\description{
Read reference data
}
