% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\docType{methods}
\name{result_path}
\alias{result_path}
\alias{result_path,gdr_test_data-method}
\title{Method result_path}
\usage{
result_path(x)

\S4method{result_path}{gdr_test_data}(x)
}
\arguments{
\item{x}{object class gdr_test_data}
}
\value{
value of slot \code{result_path}
}
\description{
Method for object gdr_test_data - access to slot \code{result_path}
}
\examples{
 td <- get_test_data()
 result_file_path <- result_path(td)

}
\keyword{internal}
\keyword{methods}
\keyword{test_data_class}
