% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{standardize_df}
\alias{standardize_df}
\title{standardize_df}
\usage{
standardize_df(df)
}
\arguments{
\item{df}{a dataframe for standardization}
}
\value{
a standardized dataframe
}
\description{
Transform all the columns in a dataframe to character type
}
