% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdr_test_data_class.R
\docType{methods}
\name{template_path}
\alias{template_path}
\alias{template_path,gdr_test_data-method}
\title{Method template_path}
\usage{
template_path(x)

\S4method{template_path}{gdr_test_data}(x)
}
\arguments{
\item{x}{object class gdr_test_data}
}
\value{
value of slot \code{template_path}
}
\description{
Method for object gdr_test_data - access to slot \code{template_path}
}
\examples{
 td <- get_test_data()
 template_file_path <- template_path(td)
 
}
\keyword{methods}
